pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "jasusi-cli"
version = "0.14.0"
description = "JasusiCLI — multi-provider AI agent for the terminal"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["ai", "coding-agent", "llm", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "httpx>=0.27",
    "anyio>=4.4",
    "click>=8.1",
    "openai>=1.30",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
memory = ["chromadb>=0.5"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.3",
    "mypy>=1.10",
    "httpx>=0.27",
    "ruff>=0.4",
]

[project.scripts]
jasusi = "jasusi_cli.cli.entry:main"

[project.urls]
Repository = "https://github.com/jasusi/jasusicli"

[tool.setuptools.packages.find]
where = ["."]
include = ["jasusi_cli*"]

[tool.pytest.ini_options]
testpaths = ["jasusi_cli/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--tb=short -q"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
exclude = ["rust/", "target/"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
